requirements.txt
transformers
translate-toolkit
pytorch_lightning